requirements.txt
# Sistema de Identificação de Dados Sensíveis
# Dependências Python

# Manipulação de dados
pandas>=1.5.0

# Leitura de Excel
openpyxl>=3.0.0

# GLiNER - Detecção de PII por Machine Learning
gliner>=0.2.0
torch>=2.0.0

# Opcional: métricas avançadas
# scikit-learn>=1.0.0

# Opcional: Parquet (big data)
# pyarrow>=10.0.0

# Opcional: validação de e-mail
# email-validator>=1.3.0
google-generativeai
